Fix CodeFactor Issues - [SA1402]
